Split apart MemoryFPlayer
