+++ src/main/java/net/redstoneore/legacyfactions/cmd/Arguments.java
@@ -14,7 +14,7 @@ import net.redstoneore.legacyfactions.entity.FPlayer;
 import net.redstoneore.legacyfactions.entity.FPlayerColl;
 import net.redstoneore.legacyfactions.entity.Faction;
 import net.redstoneore.legacyfactions.entity.FactionColl;
-import net.redstoneore.legacyfactions.entity.persist.memory.MemoryFPlayer;
+import net.redstoneore.legacyfactions.entity.persist.shared.SharedFPlayer;
 import net.redstoneore.legacyfactions.util.UUIDUtil;
 
 public class Arguments {
@@ -128,10 +128,8 @@ public class Arguments {
 						
 						FPlayer fplayer = FPlayerColl.get(uuid);
 						
-						if (fplayer instanceof MemoryFPlayer) {
-							((MemoryFPlayer)fplayer).setName(value);
-						}
-						
+						((SharedFPlayer) fplayer).setName(value);
+												
 						callback.then(Optional.of((T) fplayer), Optional.empty());
 					});
 				});
