+++ src/main/java/net/redstoneore/legacyfactions/cmd/FCommand.java
@@ -11,7 +11,7 @@ import net.redstoneore.legacyfactions.entity.FPlayer;
 import net.redstoneore.legacyfactions.entity.FPlayerColl;
 import net.redstoneore.legacyfactions.entity.Faction;
 import net.redstoneore.legacyfactions.entity.FactionColl;
-import net.redstoneore.legacyfactions.entity.persist.memory.MemoryFPlayer;
+import net.redstoneore.legacyfactions.entity.persist.shared.SharedFPlayer;
 import net.redstoneore.legacyfactions.event.EventFactionsCommandExecute;
 import net.redstoneore.legacyfactions.integration.vault.VaultEngine;
 import net.redstoneore.legacyfactions.util.TextUtil;
@@ -237,9 +237,7 @@ public abstract class FCommand extends MCommand<Factions> {
 			if (uuid != null) {
 				// For sanity sake, set the player name
 				FPlayer found = FPlayerColl.get(uuid);
-				if (found instanceof MemoryFPlayer) {
-					((MemoryFPlayer) found).setName(playerName);
-				}
+				((SharedFPlayer) found).setName(playerName);
 			}
 			
 			callback.then(uuid, Optional.empty());
@@ -258,9 +256,7 @@ public abstract class FCommand extends MCommand<Factions> {
 			
 			// For sanity sake, set the player name
 			FPlayer found = FPlayerColl.get(uuid);
-			if (found instanceof MemoryFPlayer) {
-				((MemoryFPlayer)found).setName(playerName);
-			}
+			((SharedFPlayer)found).setName(playerName);
 			
 			callback.then(found, Optional.empty());
 		});
