+++ src/main/java/net/redstoneore/legacyfactions/entity/persist/memory/MemoryFPlayer.java
@@ -1,41 +1,25 @@
 package net.redstoneore.legacyfactions.entity.persist.memory;
 
-import net.redstoneore.legacyfactions.event.EventFactionsDisband;
-import org.bukkit.*;
-import org.bukkit.entity.Player;
-import org.bukkit.inventory.ItemStack;
 
-import net.redstoneore.legacyfactions.*;
-import net.redstoneore.legacyfactions.entity.Board;
+import net.redstoneore.legacyfactions.FLocation;
+import net.redstoneore.legacyfactions.Factions;
+import net.redstoneore.legacyfactions.Lang;
+import net.redstoneore.legacyfactions.Role;
+import net.redstoneore.legacyfactions.Volatile;
 import net.redstoneore.legacyfactions.entity.Conf;
-import net.redstoneore.legacyfactions.entity.FPlayer;
-import net.redstoneore.legacyfactions.entity.FPlayerColl;
 import net.redstoneore.legacyfactions.entity.Faction;
 import net.redstoneore.legacyfactions.entity.FactionColl;
-import net.redstoneore.legacyfactions.entity.VaultAccount;
-import net.redstoneore.legacyfactions.event.EventFactionsChange;
-import net.redstoneore.legacyfactions.event.EventFactionsLandChange;
-import net.redstoneore.legacyfactions.event.EventFactionsChange.ChangeReason;
+import net.redstoneore.legacyfactions.entity.persist.shared.SharedFPlayer;
 import net.redstoneore.legacyfactions.event.EventFactionsRoleChanged;
 import net.redstoneore.legacyfactions.expansion.chat.ChatMode;
-import net.redstoneore.legacyfactions.integration.essentials.EssentialsEngine;
-import net.redstoneore.legacyfactions.integration.vault.VaultEngine;
-import net.redstoneore.legacyfactions.integration.worldguard.WorldGuardEngine;
-import net.redstoneore.legacyfactions.integration.worldguard.WorldGuardIntegration;
 import net.redstoneore.legacyfactions.locality.Locality;
-import net.redstoneore.legacyfactions.mixin.PlayerMixin;
-import net.redstoneore.legacyfactions.placeholder.FactionsPlaceholders;
-import net.redstoneore.legacyfactions.scoreboards.FScoreboards;
-import net.redstoneore.legacyfactions.scoreboards.sidebar.FInfoSidebar;
-import net.redstoneore.legacyfactions.util.RelationUtil;
-import net.redstoneore.legacyfactions.util.TextUtil;
-import net.redstoneore.legacyfactions.util.TitleUtil;
 import net.redstoneore.legacyfactions.util.WarmUpUtil;
 
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
 import java.util.UUID;
+
+import org.bukkit.Bukkit;
+import org.bukkit.OfflinePlayer;
+import org.bukkit.Statistic;
 
 
 /**
@@ -53,291 +37,23 @@ import java.util.UUID;
  * <br><br>
  * Do not store references to any fields. Always use the methods available.  
  */
-public abstract class MemoryFPlayer implements FPlayer {
+public abstract class MemoryFPlayer extends SharedFPlayer {
 	
 	// -------------------------------------------------- //
-	// FIELDS
+	// STATIC FIELDS
+	// -------------------------------------------------- //
+
+	private static Locality DEFAULT_LASTSTOODAT = Locality.of(Bukkit.getWorlds().get(0).getSpawnLocation());
+	
+	// -------------------------------------------------- //
+	// CONSTRUCT
 	// -------------------------------------------------- //
 	
-	protected String factionId = "0";
-	protected Role role = Role.NORMAL;
-	protected String title = "";
-	protected double power;
-	protected double powerBoost;
-	protected long lastPowerUpdateTime;
-	protected long lastLoginTime;
-	protected ChatMode chatMode;
-	protected boolean ignoreAllianceChat = false;
-	protected String id;
-	protected String name;
-	protected boolean monitorJoins;
-	protected boolean spyingChat = false;
-	protected boolean showScoreboard = true;
-	protected WarmUpUtil.Warmup warmup;
-	protected int warmupTask;
-	protected boolean isAdminBypassing = false;
-	protected int kills, deaths;
-	protected boolean willAutoLeave = true;
-	protected boolean territoryTitlesOff = false;
+	/**
+	 * Construct this if we're going to populate fields ourselves.
+	 */
+	public MemoryFPlayer() { }
 	
-	protected transient FLocation lastStoodAt = new FLocation("world", 0, 0); // Where did this player stand the last time we checked?
-	protected transient boolean mapAutoUpdating;
-	protected transient Faction autoClaimFor;
-	protected transient boolean autoSafeZoneEnabled;
-	protected transient boolean autoWarZoneEnabled;
-	protected transient boolean loginPvpDisabled;
-	protected transient long lastFrostwalkerMessage;
-
-
-	public void onLogin() {
-		this.kills = getPlayer().getStatistic(Statistic.PLAYER_KILLS);
-		this.deaths = getPlayer().getStatistic(Statistic.DEATHS);
-	}
-
-	public void onLogout() {
-		// Ensure power is up to date
-		this.getPower();
-		
-		// Update last login time
-		this.setLastLoginTime(System.currentTimeMillis());
-		
-		// Store statistics 
-		this.kills = getPlayer().getStatistic(Statistic.PLAYER_KILLS);
-		this.deaths = getPlayer().getStatistic(Statistic.DEATHS);
-		
-		// Remove from stuck map
-		if (Volatile.get().stuckMap().containsKey(this.getPlayer().getUniqueId())) {
-			Volatile.get().stuckMap().remove(this.getPlayer().getUniqueId());
-			Volatile.get().stuckTimers().remove(this.getPlayer().getUniqueId());
-		}
-		
-		if (!this.getFaction().isWilderness()) {
-			// Toggle
-			this.getFaction().memberLoggedOff();
-			
-			// Notify members if required
-			this.getFaction().getWhereOnline(true)
-				.stream()
-				.forEach(fplayer -> {
-					if (fplayer == this || !fplayer.isMonitoringJoins()) return;
-					
-					fplayer.sendMessage(Lang.FACTION_LOGOUT, this.getName());
-				});
-
-		}
-	}
-
-	public ItemStack getItemInMainHand() {
-		return PlayerMixin.getItemInMainHand(this.getPlayer());
-	}
-	
-	public ItemStack getItemInOffHand() {
-		return PlayerMixin.getItemInOffHand(this.getPlayer());
-	}
-	
-	@Override
-	public void teleport(Locality locality) {
-		this.getPlayer().teleport(locality.getLocation());
-	}
-	
-	public Faction getFaction() {
-		if (this.factionId == null) {
-			this.factionId = "0";
-		}
-		return FactionColl.get().getFactionById(this.factionId);
-	}
-
-	public String getFactionId() {
-		return this.factionId;
-	}
-
-	public boolean hasFaction() {
-		return !factionId.equals("0");
-	}
-
-	public void setFaction(Faction faction) {
-		Faction oldFaction = this.getFaction();
-		if (oldFaction != null) {
-			oldFaction.removeFPlayer(this);
-		}
-		faction.addFPlayer(this);
-		this.factionId = faction.getId();
-	}
-
-	public void setMonitorJoins(boolean monitor) {
-		this.monitorJoins = monitor;
-	}
-
-	public boolean isMonitoringJoins() {
-		return this.monitorJoins;
-	}
-	
-	public Role getRole() {
-		// Coleader check
-		if (this.role == Role.COLEADER && Conf.enableColeaders == false) {
-			this.role = Role.NORMAL;
-		}
-		return this.role;
-	}
-	
-	public void setRole(Role role) {
-		// Coleader check
-		if (role == Role.COLEADER && Conf.enableColeaders == false) {
-			role = Role.NORMAL;
-		}
-
-		Role previousRole = this.role;
-		this.role = role;
-		EventFactionsRoleChanged.create(this.getFaction(), this, previousRole, role).call();
-	}
-	
-	@Override
-	public boolean canAdminister(FPlayer who) {
-		if (!who.getFaction().equals(this.getFaction())) {
-			who.sendMessage(TextUtil.get().parse(Lang.COMMAND_ERRORS_NOTSAME.toString().replaceAll("<name>", this.describeTo(who, true))));
-			return false;
-		}
-
-		if (who.getRole().isMoreThan(this.getRole()) || who.getRole().equals(Role.ADMIN)) {
-			return true;
-		}
-
-		if (this.getRole().equals(Role.ADMIN)) {
-			who.sendMessage(TextUtil.get().parse(Lang.COMMAND_ERRORS_ONLYFACTIONADMIN.toString()));
-		} else if (who.getRole().equals(Role.MODERATOR)) {
-			if (who == this) return true;
-			
-			who.sendMessage(TextUtil.get().parse(Lang.COMMAND_ERRORS_MODERATORSCANT.toString()));
-		} else if (who.getRole().equals(Role.COLEADER)) {
-			if (who == this) return true;
-
-			who.sendMessage(TextUtil.get().parse(Lang.COMMAND_ERRORS_COLEADERSCANT.toString()));
-		} else {
-			who.sendMessage(TextUtil.get().parse(Lang.COMMAND_ERRORS_NOTMODERATOR.toString()));
-		}
-		
-		return false;
-	}
-
-	@Override
-	public boolean hasPermission(String permission) {
-		return this.getPlayer().hasPermission(permission);
-	}
-	
-	public double getPowerBoost() {
-		return this.powerBoost;
-	}
-
-	public void setPowerBoost(double powerBoost) {
-		this.powerBoost = powerBoost;
-	}
-
-	public boolean willAutoLeave() {
-		return this.willAutoLeave;
-	}
-
-	public void setAutoLeave(boolean willLeave) {
-		this.willAutoLeave = willLeave;
-		Factions.get().debug(name + " set autoLeave to " + willLeave);
-	}
-
-	public long getLastFrostwalkerMessage() {
-		return this.lastFrostwalkerMessage;
-	}
-
-	public void setLastFrostwalkerMessage() {
-		this.lastFrostwalkerMessage = System.currentTimeMillis();
-	}
-
-	public Faction getAutoClaimFor() {
-		return autoClaimFor;
-	}
-
-	public void setAutoClaimFor(Faction faction) {
-		this.autoClaimFor = faction;
-		if (this.autoClaimFor != null) {
-			if ( ! this.autoClaimFor.isSafeZone()) {
-				this.autoSafeZoneEnabled = false;
-			}
-			
-			if ( ! this.autoClaimFor.isWarZone()) {
-				this.autoWarZoneEnabled = false;
-			}
-		}
-	}
-
-	public boolean isAutoSafeClaimEnabled() {
-		return autoSafeZoneEnabled;
-	}
-
-	public void setIsAutoSafeClaimEnabled(boolean enabled) {
-		this.autoSafeZoneEnabled = enabled;
-		if (enabled) {
-			this.autoClaimFor = null;
-			this.autoWarZoneEnabled = false;
-		}
-	}
-
-	public boolean isAutoWarClaimEnabled() {
-		return autoWarZoneEnabled;
-	}
-
-	public void setIsAutoWarClaimEnabled(boolean enabled) {
-		this.autoWarZoneEnabled = enabled;
-		if (enabled) {
-			this.autoClaimFor = null;
-			this.autoSafeZoneEnabled = false;
-		}
-	}
-
-	public boolean isAdminBypassing() {
-		return this.isAdminBypassing;
-	}
-	
-	public boolean isVanished(FPlayer viewer) {
-		return EssentialsEngine.isVanished(this.getPlayer()) || viewer.getPlayer().canSee(this.getPlayer());
-	}
-	
-	public void setIsAdminBypassing(boolean val) {
-		this.isAdminBypassing = val;
-	}
-
-	public void setChatMode(ChatMode chatMode) {
-		this.chatMode = chatMode;
-	}
-	
-	public ChatMode getChatMode() {
-		// If we're in the wilderness or factions chat is disabled, default to public chat
-		if (this.factionId.equals("0") || !Conf.expansionsFactionsChat.enabled) {
-			this.chatMode = ChatMode.PUBLIC;
-		}
-		return this.chatMode;
-	}
-
-	public void setIgnoreAllianceChat(boolean ignore) {
-		this.ignoreAllianceChat = ignore;
-	}
-
-	public boolean isIgnoreAllianceChat() {
-		return ignoreAllianceChat;
-	}
-
-	public void setSpyingChat(boolean chatSpying) {
-		this.spyingChat = chatSpying;
-	}
-
-	public boolean isSpyingChat() {
-		return spyingChat;
-	}
-
-	// FIELD: account
-	public String getAccountId() {
-		return this.getId();
-	}
-
-	public MemoryFPlayer() {
-	}
-
 	public MemoryFPlayer(String id) {
 		this.id = id;
 		this.resetFactionData();
@@ -381,7 +97,317 @@ public abstract class MemoryFPlayer implements FPlayer {
 		this.deaths = other.deaths;
 		this.territoryTitlesOff = other.territoryTitlesOff;
 	}
+	
+	// -------------------------------------------------- //
+	// FIELDS
+	// -------------------------------------------------- //
+	
+	protected String factionId = "0";
+	protected Role role = Role.NORMAL;
+	protected String title = "";
+	protected double power;
+	protected double powerBoost;
+	protected long lastPowerUpdateTime;
+	protected long lastLoginTime;
+	protected ChatMode chatMode;
+	protected boolean ignoreAllianceChat = false;
+	protected String id;
+	protected String name;
+	protected boolean monitorJoins;
+	protected boolean spyingChat = false;
+	protected boolean showScoreboard = true;
+	protected WarmUpUtil.Warmup warmup;
+	protected int warmupTask;
+	protected boolean isAdminBypassing = false;
+	protected int kills, deaths;
+	protected boolean willAutoLeave = true;
+	protected boolean territoryTitlesOff = false;
+	
+	protected transient Locality lastStoodAt = DEFAULT_LASTSTOODAT;
+	protected transient boolean mapAutoUpdating;
+	protected transient Faction autoClaimFor;
+	protected transient boolean autoSafeZoneEnabled;
+	protected transient boolean autoWarZoneEnabled;
+	protected transient boolean loginPvpDisabled;
+	protected transient long lastFrostwalkerMessage;
 
+
+	@Override
+	public String getId() {
+		return id;
+	}
+	
+	@Override
+	public void setId(String id) {
+		this.id = id;
+	}
+	
+	// -------------------------------------------------- //
+	// Title, Name, Faction Tag and Chat
+	// -------------------------------------------------- //
+	
+	@Override
+	public String getTitle() {
+		return this.hasFaction() ? this.title : Lang.NOFACTION_PREFIX.toString();
+	}
+
+	@Override
+	public void setTitle(String title) {
+		this.title = title;
+	}
+
+	@Override
+	public String getName() {
+		if (this.name == null) {
+			OfflinePlayer offline = Bukkit.getOfflinePlayer(UUID.fromString(this.getId()));
+			this.name = offline.getName() != null ? offline.getName() : this.getId();
+		}
+		return this.name;
+	}
+	
+	@Override
+	public void setName(String name) {
+		this.name = name;
+	}
+	
+	@Override
+	public int getKills() {
+		return isOnline() ? getPlayer().getStatistic(Statistic.PLAYER_KILLS) : this.kills;
+	}
+	
+	@Override
+	public int getDeaths() {
+		return isOnline() ? getPlayer().getStatistic(Statistic.DEATHS) : this.deaths;
+	}
+
+	// -------------------------------------------------- //
+	// POWER
+	// -------------------------------------------------- //
+	
+	@Override
+	public double getPower() {
+		this.updatePower();
+		return this.power;
+	}
+
+	@Override
+	public void alterPower(double delta) {
+		this.power += delta;
+		if (this.power > this.getPowerMax()) {
+			this.power = this.getPowerMax();
+		} else if (this.power < this.getPowerMin()) {
+			this.power = this.getPowerMin();
+		}
+	}
+	
+	@Override
+	public long getLastPowerUpdated() {
+		return this.lastPowerUpdateTime;
+	}
+
+	@Override
+	public void setLastPowerUpdated(long time) {
+		this.lastPowerUpdateTime = time;
+	}
+	
+	// -------------------------------------------------- //
+	// EVENTS
+	// -------------------------------------------------- //
+
+	@Override
+	public void onLogin() {
+		this.kills = getPlayer().getStatistic(Statistic.PLAYER_KILLS);
+		this.deaths = getPlayer().getStatistic(Statistic.DEATHS);
+	}
+
+	@Override
+	public void onLogout() {
+		// Ensure power is up to date
+		this.getPower();
+		
+		// Update last login time
+		this.setLastLoginTime(System.currentTimeMillis());
+		
+		// Store statistics 
+		this.kills = getPlayer().getStatistic(Statistic.PLAYER_KILLS);
+		this.deaths = getPlayer().getStatistic(Statistic.DEATHS);
+		
+		// Remove from stuck map
+		if (Volatile.get().stuckMap().containsKey(this.getPlayer().getUniqueId())) {
+			Volatile.get().stuckMap().remove(this.getPlayer().getUniqueId());
+			Volatile.get().stuckTimers().remove(this.getPlayer().getUniqueId());
+		}
+		
+		if (!this.getFaction().isWilderness()) {
+			// Toggle
+			this.getFaction().memberLoggedOff();
+			
+			// Notify members if required
+			this.getFaction().getWhereOnline(true)
+				.stream()
+				.filter(fplayer -> fplayer == this || !fplayer.isMonitoringJoins())
+				.forEach(fplayer -> fplayer.sendMessage(Lang.FACTION_LOGOUT, this.getName()));
+		}
+	}
+	
+	@Override
+	public Faction getFaction() {
+		if (this.getFactionId() == null) {
+			this.factionId = "0";
+		}
+		return FactionColl.get().getFactionById(this.getFactionId());
+	}
+
+	@Override
+	public String getFactionId() {
+		return this.factionId;
+	}
+
+	@Override
+	public void setFaction(Faction faction) {
+		Faction oldFaction = this.getFaction();
+		if (oldFaction != null) {
+			oldFaction.removeFPlayer(this);
+		}
+		faction.addFPlayer(this);
+		this.factionId = faction.getId();
+	}
+
+	@Override
+	public void setMonitorJoins(boolean monitor) {
+		this.monitorJoins = monitor;
+	}
+
+	@Override
+	public boolean isMonitoringJoins() {
+		return this.monitorJoins;
+	}
+	
+	@Override
+	public Role getRole() {
+		// Coleader check
+		if (this.role == Role.COLEADER && Conf.enableColeaders == false) {
+			this.role = Role.NORMAL;
+		}
+		return this.role;
+	}
+	
+	@Override
+	public void setRole(Role role) {
+		// Coleader check
+		if (role == Role.COLEADER && Conf.enableColeaders == false) {
+			role = Role.NORMAL;
+		}
+
+		Role previousRole = this.role;
+		this.role = role;
+		EventFactionsRoleChanged.create(this.getFaction(), this, previousRole, role).call();
+	}
+	
+	@Override
+	public double getPowerBoost() {
+		return this.powerBoost;
+	}
+
+	@Override
+	public void setPowerBoost(double powerBoost) {
+		this.powerBoost = powerBoost;
+	}
+
+	@Override
+	public boolean willAutoLeave() {
+		return this.willAutoLeave;
+	}
+
+	@Override
+	public void setAutoLeave(boolean willLeave) {
+		this.willAutoLeave = willLeave;
+		Factions.get().debug(name + " set autoLeave to " + willLeave);
+	}
+
+	@Override
+	public long getLastFrostwalkerMessage() {
+		return this.lastFrostwalkerMessage;
+	}
+
+	@Override
+	public void setLastFrostwalkerMessage() {
+		this.lastFrostwalkerMessage = System.currentTimeMillis();
+	}
+
+	@Override
+	public Faction getAutoClaimFor() {
+		return this.autoClaimFor;
+	}
+
+	@Override
+	public void setAutoClaimFor(Faction faction) {
+		this.autoClaimFor = faction;
+		if (this.autoClaimFor != null) {
+			if (!this.autoClaimFor.isSafeZone()) {
+				this.autoSafeZoneEnabled = false;
+			}
+			
+			if (!this.autoClaimFor.isWarZone()) {
+				this.autoWarZoneEnabled = false;
+			}
+		}
+	}
+
+	@Override
+	public boolean isAutoSafeClaimEnabled() {
+		return this.autoSafeZoneEnabled;
+	}
+
+	@Override
+	public void setIsAutoSafeClaimEnabled(boolean enabled) {
+		this.autoSafeZoneEnabled = enabled;
+		if (enabled) {
+			this.autoClaimFor = null;
+			this.autoWarZoneEnabled = false;
+		}
+	}
+
+	@Override
+	public boolean isAutoWarClaimEnabled() {
+		return this.autoWarZoneEnabled;
+	}
+
+	@Override
+	public void setIsAutoWarClaimEnabled(boolean enabled) {
+		this.autoWarZoneEnabled = enabled;
+		if (enabled) {
+			this.autoClaimFor = null;
+			this.autoSafeZoneEnabled = false;
+		}
+	}
+
+	@Override
+	public boolean isAdminBypassing() {
+		return this.isAdminBypassing;
+	}
+	
+	@Override
+	public void setIsAdminBypassing(boolean val) {
+		this.isAdminBypassing = val;
+	}
+
+	@Override
+	public void setChatMode(ChatMode chatMode) {
+		this.chatMode = chatMode;
+	}
+	
+	@Override
+	public ChatMode getChatMode() {
+		// If we're in the wilderness or factions chat is disabled, default to public chat
+		if (this.factionId.equals("0") || !Conf.expansionsFactionsChat.enabled) {
+			this.chatMode = ChatMode.PUBLIC;
+		}
+		return this.chatMode;
+	}
+
+
+	@Override
 	public void resetFactionData() {
 		// clean up any territory ownership in old faction, if there is one
 		if (factionId != null && FactionColl.get().isValidFactionId(this.getFactionId())) {
@@ -398,18 +424,38 @@ public abstract class MemoryFPlayer implements FPlayer {
 		this.title = "";
 		this.autoClaimFor = null;	}
 
-	// -------------------------------------------- //
-	// Getters And Setters
-	// -------------------------------------------- //
+	// -------------------------------------------------- //
+	// GETTS AND SETTERS
+	// -------------------------------------------------- //
 
+	@Override
+	public void setIgnoreAllianceChat(boolean ignore) {
+		this.ignoreAllianceChat = ignore;
+	}
 
+	@Override
+	public boolean isIgnoreAllianceChat() {
+		return this.ignoreAllianceChat;
+	}
+	
+	@Override
+	public void setSpyingChat(boolean chatSpying) {
+		this.spyingChat = chatSpying;
+	}
+
+	@Override
+	public boolean isSpyingChat() {
+		return this.spyingChat;
+	}
+
+	@Override
 	public long getLastLoginTime() {
 		return lastLoginTime;
 	}
 
-
+	@Override
 	public void setLastLoginTime(long lastLoginTime) {
-		losePowerFromBeingOffline();
+		this.losePowerFromBeingOffline();
 		this.lastLoginTime = lastLoginTime;
 		this.lastPowerUpdateTime = lastLoginTime;
 		if (Conf.noPVPDamageToOthersForXSecondsAfterLogin > 0) {
@@ -417,16 +463,19 @@ public abstract class MemoryFPlayer implements FPlayer {
 		}
 	}
 
+	@Override
 	public boolean isMapAutoUpdating() {
-		return mapAutoUpdating;
+		return this.mapAutoUpdating;
 	}
 
+	@Override
 	public void setMapAutoUpdating(boolean mapAutoUpdating) {
 		this.mapAutoUpdating = mapAutoUpdating;
 	}
 
+	@Override
 	public boolean hasLoginPvpDisabled() {
-		if (!loginPvpDisabled) {
+		if (!this.loginPvpDisabled) {
 			return false;
 		}
 		if (this.lastLoginTime + (Conf.noPVPDamageToOthersForXSecondsAfterLogin * 1000) < System.currentTimeMillis()) {
@@ -436,328 +485,31 @@ public abstract class MemoryFPlayer implements FPlayer {
 		return true;
 	}
 
+	@Override
 	public FLocation getLastStoodAt() {
+		return FLocation.valueOf(this.lastStoodAt.getLocation());
+	}
+	
+	@Override
+	public Locality getLastLocation() {
 		return this.lastStoodAt;
 	}
-	
-	public Locality getLastLocation() {
-		try {
-			return Locality.of(this.lastStoodAt.toString());
-		} catch (Exception e) {
-			e.printStackTrace();
-		}
-		return null;
-	}
 
+	@Override
 	public void setLastStoodAt(FLocation flocation) {
-		this.lastStoodAt = flocation;
+		this.lastStoodAt = Locality.of(flocation.getChunk());
 	}
 
-	//----------------------------------------------//
-	// Title, Name, Faction Tag and Chat
-	//----------------------------------------------//
-
-	// Base:
-
-	public String getTitle() {
-		return this.hasFaction() ? this.title : Lang.NOFACTION_PREFIX.toString();
-	}
-
-	public void setTitle(String title) {
-		this.title = title;
-	}
-
-	public String getName() {
-		if (this.name == null) {
-			OfflinePlayer offline = Bukkit.getOfflinePlayer(UUID.fromString(this.getId()));
-			this.name = offline.getName() != null ? offline.getName() : this.getId();
-		}
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
-	}
-
-	public String getTag() {
-		return this.hasFaction() ? this.getFaction().getTag() : "";
-	}
-
-	// Base concatenations:
-
-	public String getNameAndSomething(String something) {
-		String ret = this.role.getPrefix();
-		if (something.length() > 0) {
-			ret += something + " ";
-		}
-		ret += this.getName();
-		return ret;
-	}
-
-	public String getNameAndTitle() {
-		return this.getNameAndSomething(this.getTitle());
-	}
-
-	public String getNameAndTag() {
-		return this.getNameAndSomething(this.getTag());
-	}
-
-	// Colored concatenations:
-	// These are used in information messages
-
-	public String getNameAndTitle(Faction faction) {
-		return this.getColorTo(faction) + this.getNameAndTitle();
-	}
-
-	public String getNameAndTitle(MemoryFPlayer fplayer) {
-		return this.getColorTo(fplayer) + this.getNameAndTitle();
-	}
-
-	// Chat Tag:
-	// These are injected into the format of global chat messages.
-
-	public String getChatTag() {
-		String format = null;
-		
-		if (this.hasFaction()) {
-			// Clone from the configuration 
-			format = Conf.expansionsFactionsChat.chatTagFormatDefault.toString();
-		} else {
-			// Clone from the configuration
-			format = Conf.expansionsFactionsChat.chatTagFormatFactionless.toString();
-		}
-		
-		// Format with Placeholders
-		format = FactionsPlaceholders.get().parse(this, format);
-		
-		return format;
-	}
-
-	// Colored Chat Tag
-	public String getChatTag(Faction faction) {
-		return this.hasFaction() ? this.getRelationTo(faction).getColor() + getChatTag() : "";
-	}
-
-	public String getChatTag(MemoryFPlayer fplayer) {
-		return this.hasFaction() ? this.getColorTo(fplayer) + getChatTag() : "";
-	}
-
-	public int getKills() {
-		return isOnline() ? getPlayer().getStatistic(Statistic.PLAYER_KILLS) : this.kills;
-	}
-
-	public int getDeaths() {
-		return isOnline() ? getPlayer().getStatistic(Statistic.DEATHS) : this.deaths;
-
-	}
-
-	// -------------------------------
-	// Relation and relation colors
-	// -------------------------------
-	
 	@Override
-	public String describe() {
-		return this.describeTo(null);
+	public boolean territoryTitlesOff() {
+		return this.territoryTitlesOff;
 	}
 	
 	@Override
-	public String describeTo(RelationParticipator that, boolean ucfirst) {
-		return RelationUtil.describeThatToMe(this, that, ucfirst);
-	}
-
-	@Override
-	public String describeTo(RelationParticipator that) {
-		return RelationUtil.describeThatToMe(this, that);
-	}
-
-	@Override
-	public Relation getRelationTo(RelationParticipator rp) {
-		return RelationUtil.getRelationTo(this, rp);
-	}
-
-	@Override
-	public Relation getRelationTo(RelationParticipator rp, boolean ignorePeaceful) {
-		return RelationUtil.getRelationTo(this, rp, ignorePeaceful);
-	}
-
-	public Relation getRelationToLocation() {
-		return Board.get().getFactionAt(this.getLastLocation()).getRelationTo(this);
-	}
-
-	@Override
-	public ChatColor getColorTo(RelationParticipator rp) {
-		return RelationUtil.getColorOfThatToMe(this, rp);
-	}
-
-	//----------------------------------------------//
-	// Health
-	//----------------------------------------------//
-	public void heal(int amnt) {
-		Player player = this.getPlayer();
-		if (player == null) {
-			return;
-		}
-		player.setHealth(player.getHealth() + amnt);
-	}
-
-
-	//----------------------------------------------//
-	// Power
-	//----------------------------------------------//
-	public double getPower() {
-		this.updatePower();
-		return this.power;
-	}
-
-	public void alterPower(double delta) {
-		this.power += delta;
-		if (this.power > this.getPowerMax()) {
-			this.power = this.getPowerMax();
-		} else if (this.power < this.getPowerMin()) {
-			this.power = this.getPowerMin();
-		}
-	}
-
-	public double getPowerMax() {
-		return Conf.powerPlayerMax + this.powerBoost;
-	}
-
-	public double getPowerMin() {
-		return Conf.powerPlayerMin + this.powerBoost;
-	}
-
-	public int getPowerRounded() {
-		return (int) Math.round(this.getPower());
-	}
-
-	public int getPowerMaxRounded() {
-		return (int) Math.round(this.getPowerMax());
-	}
-
-	public int getPowerMinRounded() {
-		return (int) Math.round(this.getPowerMin());
-	}
-
-	public void updatePower() {
-		if (this.isOffline()) {
-			losePowerFromBeingOffline();
-			if (!Conf.powerRegenOffline) {
-				return;
-			}
-		} else if (hasFaction() && getFaction().isPowerFrozen()) {
-			return; // Don't let power regen if faction power is frozen.
-		}
-		long now = System.currentTimeMillis();
-		long millisPassed = now - this.lastPowerUpdateTime;
-		this.lastPowerUpdateTime = now;
-
-		Player thisPlayer = this.getPlayer();
-		if (thisPlayer != null && thisPlayer.isDead()) {
-			return;  // don't let dead players regain power until they respawn
-		}
-
-		int millisPerMinute = 60 * 1000;
-		this.alterPower(millisPassed * Conf.powerPerMinute / millisPerMinute);
-	}
-
-	public void losePowerFromBeingOffline() {
-		if (Conf.powerOfflineLossPerDay > 0.0 && this.power > Conf.powerOfflineLossLimit) {
-			long now = System.currentTimeMillis();
-			long millisPassed = now - this.lastPowerUpdateTime;
-			this.lastPowerUpdateTime = now;
-
-			double loss = millisPassed * Conf.powerOfflineLossPerDay / (24 * 60 * 60 * 1000);
-			if (this.power - loss < Conf.powerOfflineLossLimit) {
-				loss = this.power;
-			}
-			this.alterPower(-loss);
-		}
-	}
-
-	public void onDeath() {
-		this.onDeath(Conf.powerPerDeath);
+	public void territoryTitlesOff(boolean off) {
+		this.territoryTitlesOff = off;
 	}
 	
-	public void onDeath(double powerLoss) {
-		this.updatePower();
-		this.alterPower(-powerLoss);
-		if (hasFaction()) {
-			getFaction().setLastDeath(System.currentTimeMillis());
-		}
-	}
-
-	//----------------------------------------------//
-	// Territory
-	//----------------------------------------------//
-	public boolean isInOwnTerritory() {
-		return Board.get().getFactionAt(this.getLastLocation()) == this.getFaction();
-	}
-
-	public boolean isInOthersTerritory() {
-		Faction factionHere = Board.get().getFactionAt(this.getLastLocation());
-		return factionHere != null && factionHere.isNormal() && factionHere != this.getFaction();
-	}
-
-	public boolean isInAllyTerritory() {
-		return Board.get().getFactionAt(this.getLastLocation()).getRelationTo(this).isAlly();
-	}
-
-	public boolean isInNeutralTerritory() {
-		return Board.get().getFactionAt(this.getLastLocation()).getRelationTo(this).isNeutral();
-	}
-
-	public boolean isInEnemyTerritory() {
-		return Board.get().getFactionAt(this.getLastLocation()).getRelationTo(this).isEnemy();
-	}
-
-	public void sendFactionHereMessage(Faction factionFrom) {
-		Faction factionHere = Board.get().getFactionAt(this.getLastLocation());
-		boolean showInChat = true;
-		
-		// Territory change scoreboard message
-		if (this.showInfoBoard(factionHere)) {
-			FScoreboards.get(this).setTemporarySidebar(new FInfoSidebar(factionHere));
-			showInChat = Conf.scoreboardInChat;
-		}
-		
-		// Territory change chat message
-		if (showInChat && Conf.territoryChangeText) {
-			this.sendMessage(TextUtil.get().parse(Lang.FACTION_LEAVE.format(factionFrom.getTag(this), factionHere.getTag(this))));
-		}
-		
-		// Territory change title message 
-		if (!this.territoryTitlesOff && Conf.territoryTitlesShow) {
-			String titleHeader = FactionsPlaceholders.get().parse(this.getPlayer(), Conf.territoryTitlesHeader.trim());
-			String titleFooter = FactionsPlaceholders.get().parse(this.getPlayer(), Conf.territoryTitlesFooter.trim());
-			
-			// Hide footer if needed
-			if (this.getLastLocation().getFactionHere().isWarZone() && Conf.hideFooterForWarzone) {
-				titleFooter = "";
-			}
-			
-			if (this.getLastLocation().getFactionHere().isSafeZone() && Conf.hideFooterForSafezone) {
-				titleFooter = "";
-			}
-			
-			if (this.getLastLocation().getFactionHere().isWilderness() && Conf.hideFooterForWilderness) {
-				titleFooter = "";
-			}
-			
-			TitleUtil.sendTitle(this.getPlayer(), Conf.territoryTitlesTimeFadeInTicks, Conf.territoryTitlesTimeStayTicks, Conf.territoryTitlesTimeFadeOutTicks, titleHeader, titleFooter);
-		}
-	}
-
-	/**
-	 * Check if the scoreboard should be shown. Simple method to be used by above method.
-	 *
-	 * @param toShow Faction to be shown.
-	 *
-	 * @return true if should show, otherwise false.
-	 */
-	public boolean showInfoBoard(Faction toShow) {
-		return this.showScoreboard && !toShow.isWarZone() && !toShow.isWilderness() && !toShow.isSafeZone() && !Conf.scoreboardInfo.isEmpty() && Conf.scoreboardInfoEnabled && FScoreboards.get(this) != null;
-	}
-
 	@Override
 	public boolean showScoreboard() {
 		return this.showScoreboard;
@@ -767,423 +519,7 @@ public abstract class MemoryFPlayer implements FPlayer {
 	public void setShowScoreboard(boolean show) {
 		this.showScoreboard = show;
 	}
-
-	// -------------------------------
-	// Actions
-	// -------------------------------
-
-	public void leave(boolean makePay) {
-		this.leave(makePay, false);
-	}
 	
-	public void leave(boolean makePay, boolean silent) {
-		Faction myFaction = this.getFaction();
-		makePay = makePay && VaultEngine.getUtils().shouldBeUsed() && !this.isAdminBypassing();
-
-		if (myFaction == null) {
-			resetFactionData();
-			return;
-		}
-
-		boolean perm = myFaction.isPermanent();
-		
-		if (!perm && this.getRole() == Role.ADMIN && myFaction.getMembers().size() > 1) {
-			this.sendMessage(Lang.LEAVE_PASSADMIN);
-			return;
-		}
-
-		if (!Conf.canLeaveWithNegativePower && this.getPower() < 0) {
-			this.sendMessage(Lang.LEAVE_NEGATIVEPOWER);
-			return;
-		}
-
-		// if economy is enabled and they're not on the bypass list, make sure they can pay
-		if (makePay && !VaultEngine.getUtils().hasAtLeast(this, Conf.econCostLeave, Lang.LEAVE_TOLEAVE.toString())) {
-			return;
-		}
-		
-		EventFactionsChange event = new EventFactionsChange(this, myFaction, FactionColl.get().getWilderness(), true, ChangeReason.LEAVE);
-		Bukkit.getServer().getPluginManager().callEvent(event);
-		if (event.isCancelled()) {
-			return;
-		}
-
-		// then make 'em pay (if applicable)
-		if (makePay && !VaultEngine.getUtils().modifyMoney(this, -Conf.econCostLeave, Lang.LEAVE_TOLEAVE.toString(), Lang.LEAVE_FORLEAVE.toString())) {
-			return;
-		}
-
-		// Am I the last one in the faction?
-		if (myFaction.getMembers().size() == 1) {
-			// Transfer all money
-			if (VaultEngine.getUtils().shouldBeUsed()) {
-				VaultAccount.get(myFaction).transfer(VaultAccount.get(this), VaultAccount.get(myFaction).getBalance(), VaultAccount.get(this));
-			}
-		}
-
-		if (myFaction.isNormal()) {
-			for (FPlayer fplayer : myFaction.getWhereOnline(true)) {
-				fplayer.sendMessage(Lang.LEAVE_LEFT, this.describeTo(fplayer, true), myFaction.describeTo(fplayer));
-			}
-
-			if (Conf.logFactionLeave) {
-				Factions.get().log(Lang.LEAVE_LEFT.format(this.getName(), myFaction.getTag()));
-			}
-		}
-
-		myFaction.removeAnnouncements(this);
-		this.resetFactionData();
-
-		if (myFaction.isNormal() && !perm && myFaction.getMembers().isEmpty()) {
-			EventFactionsDisband disbandEvent = new EventFactionsDisband(getPlayer(), myFaction.getId(), false,
-					EventFactionsDisband.DisbandReason.LEAVE);
-			Bukkit.getPluginManager().callEvent(disbandEvent);
-
-			// Remove this faction
-			for (FPlayer fplayer : FPlayerColl.all()) {
-				fplayer.sendMessage(Lang.LEAVE_DISBANDED, myFaction.describeTo(fplayer, true));
-			}
-
-			FactionColl.get().removeFaction(myFaction.getId());
-			if (Conf.logFactionDisband) {
-				Factions.get().log(Lang.LEAVE_DISBANDEDLOG.format(myFaction.getTag(), myFaction.getId(), this.getName()));
-			}
-		}
-	}
-
-	@Override
-	public boolean canClaimForFaction(Faction forFaction) {
-		return !forFaction.isWilderness() && (this.isAdminBypassing() || (forFaction == this.getFaction() && this.getRole().isAtLeast(Role.MODERATOR)) || (forFaction.isSafeZone() && Permission.MANAGE_SAFE_ZONE.has(getPlayer())) || (forFaction.isWarZone() && Permission.MANAGE_WAR_ZONE.has(getPlayer())));
-	}
-
-	@Override
-	public boolean canClaimForFactionAtLocation(Faction forFaction, Location location, boolean notifyFailure) {
-		FLocation flocation = new FLocation(location);
-		
-		return canClaimForFactionAtLocation(forFaction, flocation, notifyFailure);
-	}
-	
-	@Override
-	public boolean canClaimForFactionAtLocation(Faction forFaction, Locality locality, boolean notifyFailure) {
-		Faction myFaction = this.getFaction();
-		Faction currentFaction = Board.get().getFactionAt(locality);
-		int ownedLand = forFaction.getLandRounded();
-		
-		// Admin Bypass needs no further checks
-		if (this.isAdminBypassing()) return true;
-		
-		// Can claim in safe zone?
-		if (forFaction.isSafeZone() && Permission.MANAGE_SAFE_ZONE.has(getPlayer())) return true;
-		
-		// Claim for war zone
-		if (forFaction.isWarZone() && Permission.MANAGE_WAR_ZONE.has(getPlayer())) return true;
-
-		// Checks for WorldGuard regions in the chunk attempting to be claimed
-		if (WorldGuardIntegration.get().isEnabled() && Conf.worldGuardChecking && WorldGuardEngine.checkForRegionsInChunk(locality.getChunk())) {
-			this.sendMessage(notifyFailure, Lang.CLAIM_PROTECTED);
-			return false;
-		}
-		
-		// Check if this is a no-claim world
-		if (Conf.worldsNoClaiming.contains(locality.getWorld().getName())) {
-			this.sendMessage(notifyFailure, Lang.CLAIM_DISABLED);
-			return false;
-		} 
-		
-		// Can only claim for own faction
-		if (myFaction != forFaction) {
-			this.sendMessage(notifyFailure, Lang.CLAIM_CANTCLAIM, forFaction.describeTo(this));
-			return false;
-		}
-		
-		// Do we already own this?
-		if (forFaction == currentFaction) {
-			this.sendMessage(notifyFailure, Lang.CLAIM_ALREADYOWN, forFaction.describeTo(this, true));
-			return false;
-		}
-		
-		// Are they at lease a moderator?
-		if (!this.getRole().isAtLeast(Role.MODERATOR)) {
-			this.sendMessage(notifyFailure, Lang.CLAIM_MUSTBE, Role.MODERATOR.getTranslation());
-			return false;
-		}
-		
-		// Check for minimum members
-		if (forFaction.getMembers().size() < Conf.claimsRequireMinFactionMembers) {
-			this.sendMessage(notifyFailure, Lang.CLAIM_MEMBERS, Conf.claimsRequireMinFactionMembers);
-			return false;
-		}
-		
-		// Check for safezone
-		if (currentFaction.isSafeZone()) {
-			this.sendMessage(notifyFailure, Lang.CLAIM_SAFEZONE);
-			return false;
-		}
-		
-		// Check for warzone
-		if (currentFaction.isWarZone()) {
-			this.sendMessage(notifyFailure, Lang.CLAIM_WARZONE);
-			return false;
-		} 
-		
-		// Check raidable can overclaim
-		if (Conf.raidableAllowOverclaim && ownedLand >= forFaction.getPowerRounded()) {
-			this.sendMessage(notifyFailure, Lang.CLAIM_POWER);
-			return false;
-		}
-		
-		// Check for claimedLandsMax
-		if (Conf.claimedLandsMax > 0 && ownedLand >= Conf.claimedLandsMax && forFaction.isNormal()) {
-			this.sendMessage(notifyFailure, Lang.CLAIM_LIMIT);
-			return false;
-		}
-		
-		// Check for ally claim 
-		if (currentFaction.getRelationTo(forFaction) == Relation.ALLY) {
-			this.sendMessage(notifyFailure, Lang.CLAIM_ALLY.toString());
-			return false;
-		} 
-
-		// Check if must be connected
-		if (Conf.claimsMustBeConnected && !this.isAdminBypassing() && myFaction.getLandRoundedInWorld(locality.getWorld()) > 0 && !Board.get().isConnectedLocation(locality, myFaction) && (!Conf.claimsCanBeUnconnectedIfOwnedByOtherFaction || !currentFaction.isNormal())) {
-			if (Conf.claimsCanBeUnconnectedIfOwnedByOtherFaction) {
-				this.sendMessage(notifyFailure, Lang.CLAIM_CONTIGIOUS);
-			} else {
-				this.sendMessage(notifyFailure, Lang.CLAIM_FACTIONCONTIGUOUS);
-			}
-			return false;
-		}
-		
-		// Check for buffer
-		if (Conf.bufferFactions > 0 && Board.get().hasFactionWithin(locality, myFaction, Conf.bufferFactions)) {
-			this.sendMessage(notifyFailure, Lang.CLAIM_TOOCLOSETOOTHERFACTION.format(Conf.bufferFactions));
-			return false;
-		}
-		
-		// Border check
-		if (Conf.claimsCanBeOutsideBorder == false && locality.isOutsideWorldBorder(Conf.bufferWorldBorder)) {
-			if (Conf.bufferWorldBorder > 0) {
-				this.sendMessage(notifyFailure, Lang.CLAIM_OUTSIDEBORDERBUFFER.format(Conf.bufferWorldBorder));
-			} else {
-				this.sendMessage(notifyFailure, Lang.CLAIM_OUTSIDEWORLDBORDER);
-			}
-			return false;
-		}
-		
-		// If the faction we're trying to claim is a normal faction ...
-		if (currentFaction.isNormal()) {
-			
-			// .. and i'm peaceful ...
-			if (myFaction.isPeaceful()) {
-				// .. don't allow - i'm peaceful.
-				this.sendMessage(notifyFailure, Lang.CLAIM_PEACEFUL, currentFaction.getTag(this));
-				return false;
-			}
-			
-			// .. and they're peaceful ...
-			if (currentFaction.isPeaceful()) {
-				// .. don't allow - they're peaceful.
-				this.sendMessage(notifyFailure, Lang.CLAIM_PEACEFULTARGET, currentFaction.getTag(this));
-				return false;
-			}
-			
-			// .. and they're strong enough to hold it
-			if (!currentFaction.hasLandInflation()) {
-				// ... don't allow - they're too strong
-				this.sendMessage(notifyFailure, Lang.CLAIM_THISISSPARTA, currentFaction.getTag(this));
-				return false;
-			}
-			
-			// .. raidableAllowOverclaim is false, and the current faction is not strong enough 
-			if (!Conf.raidableAllowOverclaim && currentFaction.hasLandInflation()) {
-				// .. don't allow it, overclaim is disabled
-				this.sendMessage(notifyFailure, Lang.CLAIM_OVERCLAIM_DISABLED);
-				return false;
-			}
-			
-			if (!Board.get().isBorderLocation(locality)) {
-				this.sendMessage(notifyFailure, Lang.CLAIM_BORDER);
-				return false;
-			}
-		}
-		
-		// can claim!
-		return true;
-	}
-	
-	@Override
-	public boolean canClaimForFactionAtLocation(Faction forFaction, FLocation flocation, boolean notifyFailure) {
-		return this.canClaimForFactionAtLocation(forFaction, Locality.of(flocation.getChunk()), notifyFailure);
-	}
-	
-	@Override
-	public boolean attemptClaim(Faction forFaction, Locality locality, boolean notifyFailure, EventFactionsLandChange eventLandChange) {
-		return this.attemptClaim(forFaction, locality, notifyFailure, notifyFailure);
-	}
-	
-	@Override
-	public boolean attemptClaim(Faction forFaction, Location location, boolean notifyFailure, EventFactionsLandChange eventLandChange) {
-		return this.attemptClaim(forFaction, location, notifyFailure, notifyFailure);
-	}
-		
-	@Override
-	public boolean attemptClaim(Faction forFaction, Location location, boolean notifyFailure, boolean notifySuccess) {		
-		return this.attemptClaim(forFaction, Locality.of(location), notifyFailure, notifySuccess);
-	}
-	
-	@Override
-	public boolean attemptClaim(Faction forFaction, FLocation flocation, boolean notifyFailure, EventFactionsLandChange eventLandChange) {
-		return this.attemptClaim(forFaction, Locality.of(flocation.getChunk()), notifyFailure, notifyFailure);
-	}
-	
-	@Override
-	public boolean attemptClaim(Faction forFaction, Locality locality, boolean notifyFailure, boolean notifySuccess) {
-		Faction currentFaction = Board.get().getFactionAt(locality);
-		
-		int ownedLand = forFaction.getLandRounded();
-
-		if (!this.canClaimForFactionAtLocation(forFaction, locality, notifyFailure)) {
-			return false;
-		}
-
-		// if economy is enabled and they're not on the bypass list, make sure they can pay
-		boolean mustPay = VaultEngine.getUtils().shouldBeUsed() && !this.isAdminBypassing() && !forFaction.isSafeZone() && !forFaction.isWarZone();
-		double cost = 0.0;
-		EconomyParticipator payee = null;
-		if (mustPay) {
-			cost = VaultEngine.getUtils().calculateClaimCost(ownedLand, currentFaction.isNormal(), locality);
-
-			if (Conf.econClaimUnconnectedFee != 0.0 && forFaction.getLandRoundedInWorld(locality.getWorld()) > 0 && !Board.get().isConnectedLocation(locality, forFaction)) {
-				cost += Conf.econClaimUnconnectedFee;
-			}
-
-			if (Conf.bankEnabled && Conf.bankFactionPaysLandCosts && this.hasFaction()) {
-				payee = this.getFaction();
-			} else {
-				payee = this;
-			}
-			
-			if (!payee.getVaultAccount().has(cost)) {
-				payee.sendMessage(Lang.CLAIM_TOCLAIM.getBuilder().parse().toString());
-				return false;
-			}
-			
-			// then make 'em pay (if applicable)
-			if (!VaultEngine.getUtils().modifyMoney(payee, -cost, Lang.CLAIM_TOCLAIM.toString(), Lang.CLAIM_FORCLAIM.toString())) {
-				return false;
-			}
-		}
-
-		// Was an over claim
-		if (currentFaction.isNormal() && currentFaction.hasLandInflation()) {
-			// Give them money for over claiming.
-			VaultEngine.getUtils().modifyMoney(payee, Conf.econOverclaimRewardMultiplier, Lang.CLAIM_TOOVERCLAIM.toString(), Lang.CLAIM_FOROVERCLAIM.toString());
-		}
-
-		// announce success
-		if (notifySuccess) {
-			Set<FPlayer> informTheseFPlayers = new HashSet<FPlayer>();
-			informTheseFPlayers.add(this);
-			informTheseFPlayers.addAll(forFaction.getWhereOnline(true));
-			
-			informTheseFPlayers.forEach(fplayer -> fplayer.sendMessage(Lang.CLAIM_CLAIMED, this.describeTo(fplayer, true), forFaction.describeTo(fplayer), currentFaction.describeTo(fplayer)));
-		}
-		
-		Board.get().setFactionAt(forFaction, locality);
-
-		if (Conf.logLandClaims) {
-			Factions.get().log(Lang.CLAIM_CLAIMEDLOG.toString(), this.getName(), locality.getCoordString(), forFaction.getTag());
-		}
-
-		return true;
-	}
-
-	public boolean shouldBeSaved() {
-		if (!this.hasFaction() && (this.getPowerRounded() == this.getPowerMaxRounded() || this.getPowerRounded() == (int) Math.round(Conf.powerPlayerStarting))) {
-			return false;
-		}
-		return true;
-	}
-	
-	public void sendMessage(boolean onlyIfTrue, String str, Object... args) {
-		if (onlyIfTrue) this.sendMessage(str, args);
-	}
-
-	public void sendMessage(String str, Object... args) {
-		this.sendMessage(TextUtil.get().parse(str, args));
-	}
-
-	public void sendMessage(boolean onlyIfTrue, Lang translation, Object... args) {
-		if(onlyIfTrue) this.sendMessage(translation, args);
-	}
-	
-	public void sendMessage(Lang translation, Object... args) {
-		this.sendMessage(translation.toString(), args);
-	}
-	
-	public Player getPlayer() {
-		return Bukkit.getPlayer(UUID.fromString(this.getId()));
-	}
-
-	public boolean isOnline() {
-		return this.getPlayer() != null;
-	}
-
-	// make sure target player should be able to detect that this player is online
-	public boolean isOnlineAndVisibleTo(Player player) {
-		Player target = this.getPlayer();
-		return target != null && player.canSee(target);
-	}
-
-	public boolean isOffline() {
-		return !isOnline();
-	}
-
-	// -------------------------------------------- //
-	// Message Sending Helpers
-	// -------------------------------------------- //
-
-	public void sendMessage(String message) {
-		if (message.contains("{null}")) return; // user wants this message to not send
-		
-		if (message.contains("/n/")) {
-			for (String line : message.split("/n/")) {
-				this.sendMessage(line);
-			}
-			return;
-		}
-		
-		Player player = this.getPlayer();
-		if (player == null) return; 
-		
-		player.sendMessage(message);
-	}
-
-	public void sendMessage(List<String> messages) {
-		messages.forEach(message -> this.sendMessage(message));
-	}
-
-	public String getNameAndTitle(FPlayer fplayer) {
-		return this.getColorTo(fplayer) + this.getNameAndTitle();
-	}
-
-	@Override
-	public String getChatTag(FPlayer fplayer) {
-		return this.hasFaction() ? this.getRelationTo(fplayer).getColor() + getChatTag() : "";
-	}
-
-	@Override
-	public String getId() {
-		return id;
-	}
-
-	public abstract void remove();
-
-	@Override
-	public void setId(String id) {
-		this.id = id;
-	}
-
 	@Override
 	public void clearWarmup() {
 		if (warmup != null) {
@@ -1214,84 +550,6 @@ public abstract class MemoryFPlayer implements FPlayer {
 		}
 		this.warmup = warmup;
 		this.warmupTask = taskId;
-	}
-	
-	@Override
-	public MemoryFPlayer asMemoryFPlayer() {
-		return (MemoryFPlayer) this;
-	}
-	
-	// -------------------------------------------------- //
-	// UTILS
-	// -------------------------------------------------- //
-	
-	public boolean payForCommand(double cost, String toDoThis, String forDoingThis) {
-		if (!VaultEngine.getUtils().shouldBeUsed() || cost == 0.0 || this.isAdminBypassing()) {
-			return true;
-		}
-
-		if (Conf.bankEnabled && Conf.bankFactionPaysCosts && this.hasFaction()) {
-			return VaultEngine.getUtils().modifyMoney(this.getFaction(), -cost, toDoThis, forDoingThis);
-		} else {
-			return VaultEngine.getUtils().modifyMoney(this, -cost, toDoThis, forDoingThis);
-		}
-	}
-	
-	// like above, but just make sure they can pay; returns true unless person can't afford the cost
-	public boolean canAffordCommand(double cost, String toDoThis) {
-		if (!VaultEngine.getUtils().shouldBeUsed() || cost == 0.0 || this.isAdminBypassing()) {
-			return true;
-		}
-
-		if (Conf.bankEnabled && Conf.bankFactionPaysCosts && this.hasFaction()) {
-			return VaultEngine.getUtils().hasAtLeast(this.getFaction(), cost, toDoThis);
-		} else {
-			return VaultEngine.getUtils().hasAtLeast(this, cost, toDoThis);
-		}
-	}
-
-	// -------------------------------------------------- //
-	// DEPRECATED
-	// -------------------------------------------------- //
-	
-	/**
-	 * Use isVanished(FPlayer viewer)
-	 */
-	@Deprecated
-	public boolean isVanished() {
-		return EssentialsEngine.isVanished(this.getPlayer());
-	}
-	
-	/**
-	 * use sendMessage
-	 */
-	@Deprecated
-	public void msg(boolean onlyIfTrue, String str, Object... args) {
-		this.sendMessage(true, str, args);
-	}
-	
-	/**
-	 * use sendMessage
-	 */
-	@Deprecated
-	public void msg(String str, Object... args) {
-		this.sendMessage(str, args);
-	}
-
-	/**
-	 * use sendMessage
-	 */
-	@Deprecated
-	public void msg(boolean onlyIfTrue, Lang translation, Object... args) {
-		this.sendMessage(onlyIfTrue, translation, args);
-	}
-	
-	/**
-	 * use sendMessage
-	 */
-	@Deprecated
-	public void msg(Lang translation, Object... args) {
-		this.sendMessage(translation, args);
 	}
 	
 }
